Upgrade Electron to v14. Replace Spectron with Playwright. (#1985)

+++ scripts/clean.js
@@ -1,10 +1,15 @@
 const fs = require("fs-extra")
 const glob = require("glob")
 const {bold} = require("chalk")
+const flags = require("./util/flags")
 
 console.log(bold("Cleaning built files"))
-fs.removeSync("dist")
-fs.removeSync("zealot/dist")
+if (!flags.keepDist) {
+  fs.removeSync("zealot/dist")
+  fs.removeSync("dist")
+} else {
+  console.log(bold("Keeping dist directories"))
+}
 // These are broken symlinks that electron leaves around when it
 // forcefully exists and can't clean them up. Babel tries to call
 // stat on the broken symlinks and blows up.
